requirements.txt
amqp==5.0.2
arabic-reshaper==2.1.1
asgiref==3.3.1
bcrypt==3.2.0
billiard==3.6.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
cssselect2==0.4.1
deepdiff==5.2.1
defusedxml==0.6.0
Django==3.1.5
django-background-task==0.1.8
elasticsearch==7.10.1
flatten-json==0.1.7
future==0.18.2
html5lib==1.1
idna==2.10
kombu==5.0.2
numpy==1.19.5
ordered-set==4.0.2
pandas==1.2.0
paramiko==2.7.2
Pillow==8.1.0
prompt-toolkit==3.0.10
pycparser==2.20
PyNaCl==1.4.0
PyPDF2==1.26.0
Pyphen==0.10.0
pypsexec==0.2.0
pysftp==0.2.9
pyspnego==0.1.4
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.5
pyvmomi==7.0.1
reportlab==3.5.59
requests==2.25.1
scp==0.13.3
six==1.15.0
smbprotocol==1.2.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
xhtml2pdf==0.2.5
djangorestframework
parallel-ssh
TestLink-API-Python-client
django-datatables-view
pywbem
